move common dependencies to common-lib
