+++ .gitignore
@@ -12,3 +12,4 @@ node_modules
 **/public/build
 /RUNNING_PID
 .DS_Store
+.bsp
