+++ apps/checker/app/model/MatcherResponse.scala
@@ -1,5 +1,6 @@
 package model
 
+import com.gu.typerighter.model.{RuleMatch, TextBlock}
 import play.api.libs.json._
 
 case class MatcherResponse(blocks: List[TextBlock], categoryIds: Set[String], matches: List[RuleMatch]) {
