+++ apps/checker/test/scala/model/TextBlockTest.scala
@@ -1,10 +1,9 @@
 package scala.model
 
+import com.gu.typerighter.model.{Text, TextBlock, TextRange}
 import model._
 import org.scalatest.flatspec.AsyncFlatSpec
 import org.scalatest.matchers.should.Matchers
-
-import utils.Text
 
 class TextBlockTest extends AsyncFlatSpec with Matchers {
   behavior of "removeSkippedRanges"
