+++ apps/common-lib/src/main/scala/com/gu/typerighter/model/BaseRule.scala
@@ -1,4 +1,4 @@
-package model
+package com.gu.typerighter.model
 
 import java.util.{List => JList}
 import java.util.regex.Pattern
@@ -10,8 +10,7 @@ import org.languagetool.rules.{Rule => LanguageToolRule}
 
 import scala.util.matching.Regex
 import scala.jdk.CollectionConverters._
-import utils.Text
-import matchers.RegexMatcher
+
 
 /**
   * A rule to match text against.
@@ -67,7 +66,6 @@ case class RegexRule(
       replacement = transformedReplacement,
       markAsCorrect = transformedReplacement.map(_.text).getOrElse("") == block.text.substring(start, end),
       matchContext = Text.getMatchTextSnippet(precedingText, matchedText, subsequentText),
-      matcherType = RegexMatcher.getType()
     )
   }
 }
