+++ apps/common-lib/src/main/scala/com/gu/typerighter/model/PatternToken.scala
@@ -1,7 +1,7 @@
-package model
+package com.gu.typerighter.model
 
-import org.languagetool.rules.patterns.{PatternToken => LTPatternToken}
 import play.api.libs.json.{Json, Reads, Writes}
+import org.languagetool.rules.patterns.{PatternRule => LTPatternRule, PatternToken => LTPatternToken}
 
 /**
   * The application's representation of a LanguageTool PatternToken, for display purposes only.
