+++ apps/common-lib/src/main/scala/com/gu/typerighter/model/RuleMatch.scala
@@ -1,15 +1,12 @@
-package model
+package com.gu.typerighter.model
 
 import org.languagetool.rules.{RuleMatch => LTRuleMatch}
 import play.api.libs.json.{Json, Writes}
 
 import scala.jdk.CollectionConverters._
-import scala.util.matching.Regex
-
-import utils.Text
 
 object RuleMatch {
-  def fromLT(lt: LTRuleMatch, block: TextBlock, matcherType: String): RuleMatch = {
+  def fromLT(lt: LTRuleMatch, block: TextBlock): RuleMatch = {
     val suggestions = lt.getSuggestedReplacements.asScala.toList.map {
       TextSuggestion(_)
     }
@@ -30,7 +27,6 @@ object RuleMatch {
       replacement = replacement,
       suggestions = suggestions,
       matchContext = Text.getMatchTextSnippet(precedingText, matchedText, subsequentText),
-      matcherType = matcherType
     )
   }
 
@@ -49,7 +45,7 @@ case class RuleMatch(rule: BaseRule,
                      replacement: Option[Suggestion] = None,
                      markAsCorrect: Boolean = false,
                      matchContext: String,
-                     matcherType: String) {
+                     ) {
   /**
     * Map the range this match applies to through the given ranges, adjusting its range accordingly.
     */
