+++ apps/common-lib/src/main/scala/com/gu/typerighter/model/Text.scala
@@ -1,6 +1,4 @@
-package utils
-
-import model.{RuleMatch, TextBlock, TextRange}
+package com.gu.typerighter.model
 
 object Text {
   def getSurroundingText(text: String, from: Int, to: Int): (String, String) = {
@@ -18,4 +16,3 @@ object Text {
   def getMatchTextSnippet(before: String, matchedText: String, after: String) =
     before + "[[" + matchedText + "]]" + after
 }
-
