+++ apps/common-lib/src/main/scala/com/gu/typerighter/model/TextBlock.scala
@@ -1,4 +1,4 @@
-package model
+package com.gu.typerighter.model
 
 import play.api.libs.json.{Json, Reads, Writes}
