+++ apps/common-lib/src/main/scala/com/gu/typerighter/rules/SheetsRuleManager.scala
@@ -1,24 +1,18 @@
-package rules
+package com.gu.typerighter.rules
+
+import com.gu.typerighter.model.{BaseRule, Category, LTRuleXML, RegexRule, RuleResource, TextSuggestion}
+import play.api.Logging
 
 import java.io._
 import java.util.Collections
-
 import com.google.api.client.googleapis.auth.oauth2.GoogleCredential
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.services.sheets.v4.{Sheets, SheetsScopes}
-import model.{Category, RegexRule, TextSuggestion}
 
 import scala.jdk.CollectionConverters._
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
-import model.{LTRule, BaseRule, RuleResource}
-import scala.concurrent.ExecutionContext
-import services.MatcherPool
-import matchers.RegexMatcher
-import matchers.LanguageToolFactory
-import play.api.Logging
-import model.LTRuleXML
 
 object PatternRuleCols {
   val Type = 0
@@ -37,7 +31,7 @@ object PatternRuleCols {
   * @param credentialsJson A string containing the JSON the Google credentials service expects
   * @param spreadsheetId Available in the sheet URL
   */
-class SheetsRuleManager(credentialsJson: String, spreadsheetId: String, matcherPool: MatcherPool, languageToolFactory: LanguageToolFactory) extends Logging {
+class SheetsRuleManager(credentialsJson: String, spreadsheetId: String) extends Logging {
   private val APPLICATION_NAME = "Typerighter"
   private val JSON_FACTORY = JacksonFactory.getDefaultInstance
