+++ build.sbt
@@ -48,6 +48,11 @@ val commonSettings = Seq(
     "com.gu" % "kinesis-logback-appender" % "2.1.0",
     "com.gu" %% "simple-configuration-ssm" % "1.5.7",
     "com.gu" %% "pan-domain-auth-verification" % "1.2.0",
+    "com.google.api-client" % "google-api-client" % "1.23.0",
+    "com.google.oauth-client" % "google-oauth-client-jetty" % "1.23.0",
+    "com.google.apis" % "google-api-services-sheets" % "v4-rev516-1.23.0",
+    "org.languagetool" % "languagetool-core" % languageToolVersion,
+    "org.languagetool" % "language-en" % languageToolVersion,
   ),
   dependencyOverrides ++= Seq(
     "com.fasterxml.jackson.core" % "jackson-databind" % "2.11.4",
@@ -76,16 +81,11 @@ val checker = (project in file(s"$appsFolder/checker"))
     commonSettings,
     libraryDependencies ++= Seq(
       ws,
-      "org.languagetool" % "languagetool-core" % languageToolVersion,
-      "org.languagetool" % "language-en" % languageToolVersion,
       "com.amazonaws" % "aws-java-sdk-ec2" % awsSdkVersion,
       "com.amazonaws" % "aws-java-sdk-s3" % awsSdkVersion,
       "com.amazonaws" % "aws-java-sdk-ssm" % awsSdkVersion,
       "com.amazonaws" % "aws-java-sdk-kinesis" % awsSdkVersion,
       "com.amazonaws" % "aws-java-sdk-cloudwatch" % awsSdkVersion,
-      "com.google.api-client" % "google-api-client" % "1.23.0",
-      "com.google.oauth-client" % "google-oauth-client-jetty" % "1.23.0",
-      "com.google.apis" % "google-api-services-sheets" % "v4-rev516-1.23.0",
       "net.logstash.logback" % "logstash-logback-encoder" % "6.0",
       "org.webjars" % "bootstrap" % "4.3.1",
       "com.gu" %% "content-api-models-scala" % capiModelsVersion,
